Reorganize classes and variable in a more scala-idiomatic way
