--- wvlet-ui-main/src/main/scala/wvlet/lang/ui/editor/Constant.scala
@@ -1,101 +1,0 @@
-package wvlet.lang.ui.editor
-
-import scala.scalajs.js
-
-val Keywords = js.Array(
-  "test",
-  "should",
-  "be",
-  "contain",
-  "debug",
-  "def",
-  "inline",
-  "type",
-  "extends",
-  "native",
-  "show",
-  "sample",
-  "this",
-  "of",
-  "in",
-  "by",
-  "as",
-  "to",
-  "with",
-  "from",
-  "agg",
-  "select",
-  "for",
-  "let",
-  "where",
-  "group",
-  "having",
-  "order",
-  "limit",
-  "transform",
-  "pivot",
-  "distinct",
-  "asc",
-  "desc",
-  "join",
-  "on",
-  "left",
-  "right",
-  "full",
-  "inner",
-  "cross",
-  "add",
-  "exclude",
-  "rename",
-  "shift",
-  "drop",
-  "describe",
-  "concat",
-  "dedup",
-  "intersect",
-  "except",
-  "all",
-  "over",
-  "partition",
-  "unbounded",
-  "preceding",
-  "following",
-  "current",
-  "range",
-  "row",
-  "run",
-  "import",
-  "export",
-  "package",
-  "model",
-  "execute",
-  "val",
-  "if",
-  "then",
-  "else",
-  "end",
-  "and",
-  "or",
-  "not",
-  "is",
-  "like",
-  "save",
-  "append",
-  "delete",
-  "truncate"
-)
-
-val TypeKeywords = js.Array(
-  "boolean",
-  "double",
-  "byte",
-  "int",
-  "short",
-  "char",
-  "void",
-  "long",
-  "float",
-  "string",
-  "array",
-  "map"
-)
